actions: Split out zerver.actions.default_streams.
