+++ zerver/lib/actions.py
@@ -48,6 +48,11 @@ from confirmation.models import (
     create_confirmation_link,
     generate_key,
 )
+from zerver.actions.default_streams import (
+    do_remove_default_stream,
+    do_remove_streams_from_default_stream_group,
+    get_default_streams_for_realm,
+)
 from zerver.actions.user_groups import (
     do_send_user_group_members_update_event,
     update_users_in_full_members_system_group,
@@ -264,7 +269,6 @@ from zerver.models import (
     get_bot_dicts_in_realm,
     get_bot_services,
     get_client,
-    get_default_stream_groups,
     get_fake_email_domain,
     get_huddle_user_ids,
     get_old_unclaimed_attachments,
@@ -2928,24 +2932,6 @@ def check_update_message(
     return number_changed
 
 
-def check_default_stream_group_name(group_name: str) -> None:
-    if group_name.strip() == "":
-        raise JsonableError(_("Invalid default stream group name '{}'").format(group_name))
-    if len(group_name) > DefaultStreamGroup.MAX_NAME_LENGTH:
-        raise JsonableError(
-            _("Default stream group name too long (limit: {} characters)").format(
-                DefaultStreamGroup.MAX_NAME_LENGTH,
-            )
-        )
-    for i in group_name:
-        if ord(i) == 0:
-            raise JsonableError(
-                _("Default stream group name '{}' contains NULL (0x00) characters.").format(
-                    group_name,
-                )
-            )
-
-
 def send_rate_limited_pm_notification_to_bot_owner(
     sender: UserProfile, realm: Realm, content: str
 ) -> None:
@@ -5510,168 +5496,10 @@ def do_change_user_setting(
         Draft.objects.filter(user_profile=user_profile).delete()
 
 
-def lookup_default_stream_groups(
-    default_stream_group_names: List[str], realm: Realm
-) -> List[DefaultStreamGroup]:
-    default_stream_groups = []
-    for group_name in default_stream_group_names:
-        try:
-            default_stream_group = DefaultStreamGroup.objects.get(name=group_name, realm=realm)
-        except DefaultStreamGroup.DoesNotExist:
-            raise JsonableError(_("Invalid default stream group {}").format(group_name))
-        default_stream_groups.append(default_stream_group)
-    return default_stream_groups
-
-
-def notify_default_streams(realm: Realm) -> None:
-    event = dict(
-        type="default_streams",
-        default_streams=streams_to_dicts_sorted(get_default_streams_for_realm(realm.id)),
-    )
-    transaction.on_commit(lambda: send_event(realm, event, active_non_guest_user_ids(realm.id)))
-
-
-def notify_default_stream_groups(realm: Realm) -> None:
-    event = dict(
-        type="default_stream_groups",
-        default_stream_groups=default_stream_groups_to_dicts_sorted(
-            get_default_stream_groups(realm)
-        ),
-    )
-    transaction.on_commit(lambda: send_event(realm, event, active_non_guest_user_ids(realm.id)))
-
-
-def do_add_default_stream(stream: Stream) -> None:
-    realm_id = stream.realm_id
-    stream_id = stream.id
-    if not DefaultStream.objects.filter(realm_id=realm_id, stream_id=stream_id).exists():
-        DefaultStream.objects.create(realm_id=realm_id, stream_id=stream_id)
-        notify_default_streams(stream.realm)
-
-
-@transaction.atomic(savepoint=False)
-def do_remove_default_stream(stream: Stream) -> None:
-    realm_id = stream.realm_id
-    stream_id = stream.id
-    DefaultStream.objects.filter(realm_id=realm_id, stream_id=stream_id).delete()
-    notify_default_streams(stream.realm)
-
-
-def do_create_default_stream_group(
-    realm: Realm, group_name: str, description: str, streams: List[Stream]
-) -> None:
-    default_streams = get_default_streams_for_realm(realm.id)
-    for stream in streams:
-        if stream in default_streams:
-            raise JsonableError(
-                _(
-                    "'{stream_name}' is a default stream and cannot be added to '{group_name}'",
-                ).format(stream_name=stream.name, group_name=group_name)
-            )
-
-    check_default_stream_group_name(group_name)
-    (group, created) = DefaultStreamGroup.objects.get_or_create(
-        name=group_name, realm=realm, description=description
-    )
-    if not created:
-        raise JsonableError(
-            _(
-                "Default stream group '{group_name}' already exists",
-            ).format(group_name=group_name)
-        )
-
-    group.streams.set(streams)
-    notify_default_stream_groups(realm)
-
-
-def do_add_streams_to_default_stream_group(
-    realm: Realm, group: DefaultStreamGroup, streams: List[Stream]
-) -> None:
-    default_streams = get_default_streams_for_realm(realm.id)
-    for stream in streams:
-        if stream in default_streams:
-            raise JsonableError(
-                _(
-                    "'{stream_name}' is a default stream and cannot be added to '{group_name}'",
-                ).format(stream_name=stream.name, group_name=group.name)
-            )
-        if stream in group.streams.all():
-            raise JsonableError(
-                _(
-                    "Stream '{stream_name}' is already present in default stream group '{group_name}'",
-                ).format(stream_name=stream.name, group_name=group.name)
-            )
-        group.streams.add(stream)
-
-    group.save()
-    notify_default_stream_groups(realm)
-
-
-def do_remove_streams_from_default_stream_group(
-    realm: Realm, group: DefaultStreamGroup, streams: List[Stream]
-) -> None:
-    for stream in streams:
-        if stream not in group.streams.all():
-            raise JsonableError(
-                _(
-                    "Stream '{stream_name}' is not present in default stream group '{group_name}'",
-                ).format(stream_name=stream.name, group_name=group.name)
-            )
-        group.streams.remove(stream)
-
-    group.save()
-    notify_default_stream_groups(realm)
-
-
-def do_change_default_stream_group_name(
-    realm: Realm, group: DefaultStreamGroup, new_group_name: str
-) -> None:
-    if group.name == new_group_name:
-        raise JsonableError(
-            _("This default stream group is already named '{}'").format(new_group_name)
-        )
-
-    if DefaultStreamGroup.objects.filter(name=new_group_name, realm=realm).exists():
-        raise JsonableError(_("Default stream group '{}' already exists").format(new_group_name))
-
-    group.name = new_group_name
-    group.save()
-    notify_default_stream_groups(realm)
-
-
-def do_change_default_stream_group_description(
-    realm: Realm, group: DefaultStreamGroup, new_description: str
-) -> None:
-    group.description = new_description
-    group.save()
-    notify_default_stream_groups(realm)
-
-
-def do_remove_default_stream_group(realm: Realm, group: DefaultStreamGroup) -> None:
-    group.delete()
-    notify_default_stream_groups(realm)
-
-
-def get_default_streams_for_realm(realm_id: int) -> List[Stream]:
-    return [
-        default.stream
-        for default in DefaultStream.objects.select_related().filter(realm_id=realm_id)
-    ]
-
-
 def get_default_subs(user_profile: UserProfile) -> List[Stream]:
     # Right now default streams are realm-wide.  This wrapper gives us flexibility
     # to some day further customize how we set up default streams for new users.
     return get_default_streams_for_realm(user_profile.realm_id)
-
-
-# returns default streams in JSON serializable format
-def streams_to_dicts_sorted(streams: List[Stream]) -> List[Dict[str, Any]]:
-    return sorted((stream.to_dict() for stream in streams), key=lambda elt: elt["name"])
-
-
-def default_stream_groups_to_dicts_sorted(groups: List[DefaultStreamGroup]) -> List[Dict[str, Any]]:
-    return sorted((group.to_dict() for group in groups), key=lambda elt: elt["name"])
 
 
 def do_update_user_activity_interval(
