+++ zerver/tests/test_subs.py
@@ -11,33 +11,35 @@ from django.core.exceptions import ValidationError
 from django.http import HttpResponse
 from django.utils.timezone import now as timezone_now
 
-from zerver.lib.actions import (
-    bulk_add_subscriptions,
-    bulk_get_subscriber_user_ids,
-    bulk_remove_subscriptions,
+from zerver.actions.default_streams import (
     do_add_default_stream,
     do_add_streams_to_default_stream_group,
     do_change_default_stream_group_description,
     do_change_default_stream_group_name,
+    do_create_default_stream_group,
+    do_remove_default_stream,
+    do_remove_default_stream_group,
+    do_remove_streams_from_default_stream_group,
+    get_default_streams_for_realm,
+    lookup_default_stream_groups,
+)
+from zerver.lib.actions import (
+    bulk_add_subscriptions,
+    bulk_get_subscriber_user_ids,
+    bulk_remove_subscriptions,
     do_change_realm_plan_type,
     do_change_stream_post_policy,
     do_change_subscription_property,
     do_change_user_role,
-    do_create_default_stream_group,
     do_create_realm,
     do_deactivate_stream,
     do_deactivate_user,
     do_get_streams,
-    do_remove_default_stream,
-    do_remove_default_stream_group,
-    do_remove_streams_from_default_stream_group,
     do_set_realm_property,
     ensure_stream,
     gather_subscriptions,
     gather_subscriptions_helper,
-    get_default_streams_for_realm,
     get_topic_messages,
-    lookup_default_stream_groups,
     validate_user_access_to_subscribers_helper,
 )
 from zerver.lib.exceptions import JsonableError
